Broken Commit
This current update will not work.

But, I'm changing my approach with the +10s idea, I think using buttons c/d on the remote and building a flow rate up/down feature is more useful than adding time.

There's still going to be a 200s in 300s limit, or something like that, but on/off can take care of that.